--- cuda/lib/euclidean_distance.cpp
@@ -1,10 +1,0 @@
-#include <cmath>
-using namespace std;
-
-float euclidean_distance(vector<float> x, vector<float> y) {
-  float sum = 0;
-  for (int i=0; i < x.size(); i++) {
-    sum += pow(x[i] - y[i], 2);
-  }
-  return sqrt(sum);
-}
