--- cuda/lib/sequential/README.md
@@ -1,1 +1,0 @@
-## Sequential Implementation Lives Here
