--- cuda/lib/sequential/main0.cpp
@@ -1,6 +1,0 @@
-#include "./k_means_cluster.cpp"
-using namespace std;
-
-int main(){
-    
-}
