+++ lib/types.cpp
@@ -1,3 +1,6 @@
+#ifndef _TYPES
+#define _TYPES
+
 #include <unordered_map>
 #include <string>
 #include <vector>
@@ -6,3 +9,12 @@ using namespace std;
 typedef unordered_map< string, vector<float> > data_map;
 typedef vector<vector<float>> centroid_vector;
 typedef vector< vector<string> > cluster_vector;
+
+typedef struct college_dataset{
+    int size;
+    int dimensions;
+    string * names;
+    float ** features; 
+} college_dataset;
+
+#endif
