+++ stl/lib/euclidean_distance.cpp
