+++ stl/lib/helpers.cpp
