+++ stl/lib/types.cpp
@@ -1,0 +1,8 @@
+#include <unordered_map>
+#include <string>
+#include <vector>
+using namespace std;
+
+typedef unordered_map< string, vector<float> > data_map;
+typedef vector<vector<float>> centroid_vector;
+typedef vector< vector<string> > cluster_vector;
